refactor(child-pi): H-7 step 2 — extract kill/drain path to child-pi-kill.ts
Decomposed the kill-related code into a dedicated module:
- appendBoundedTail, clearHardKillTimer, spawnTaskkillSafe
- killProcessPid (exported), killProcessTree (internal), terminateActiveChildPiProcesses (exported)
- Module-level state: activeChildProcesses Map, childHardKillTimers Map, MAX_CAPTURE_BYTES
- Periodic zombie reaper (setInterval, unref'd)

Internal accessors (registerActiveChild, unregisterActiveChild,
clearHardKillTimer) exposed for runChildPi to manage the bookkeeping.
Public API (killProcessPid, terminateActiveChildPiProcesses) preserved
via re-exports.

child-pi.ts: 1672 → 1552 lines (-120). New child-pi-kill.ts: 214 lines.
All 78 child-pi + lock + dependent tests pass (no regressions).
test:bundle confirms the bundle still loads correctly.
